refactor: remove StripeService
